MediScanAI-Backend/requirements.txt
# fastapi
# uvicorn
# python-multipart
# sqlalchemy
# psycopg2-binary
# passlib[bcrypt]
# python-jose[cryptography]
# python-dotenv
# pillow
# tensorflow
# numpy


fastapi
uvicorn
tensorflow
numpy
pillow
requests
python-multipart
